# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Sistema de Monitoreo Ambiental - API",
    "version": "1.0",
    "description": "Documentación de la API del sistema de monitoreo ambiental."
  },
  "servers": [
    {
      "url": "{backendUrl}",
      "description": "Servidor HTTP del Sistema de Monitoreo Ambiental. Sustituye {backendUrl} con la URL definida en la variable de entorno BACKEND_URL para apuntar al entorno correspondiente.",
      "variables": {
        "backendUrl": {
          "default": "http://localhost:3000",
          "enum": [
            "http://localhost:3000",
            "https://monitoring-system-opbd.onrender.com",
            "http://ec2-98-86-100-220.compute-1.amazonaws.com:3000"
          ],
          "description": "Valores soportados por BACKEND_URL (local, Render y EC2)."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/reports/hourly": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Consulta los promedios horarios ya agregados",
        "description": "Retorna los registros pre-calculados por el job de agregación, incluyendo valores promedio, mínimo, máximo, número de muestras y unidad. Permite filtrar por deviceId, sensorType, día puntual o rango [from, to) y admite paginación. Es la fuente principal para graficar series horarias.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Identificador del dispositivo a filtrar."
          },
          {
            "name": "sensorType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "temperature",
                "humidity",
                "soil_humidity",
                "solar_radiation"
              ]
            },
            "description": "Tipo de sensor a filtrar."
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "description": "Fecha (YYYY-MM-DD) para recuperar las 24 horas del día."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Fecha/hora inicial (ISO 8601). Debe usarse junto con 'to'."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Fecha/hora final (ISO 8601). Debe usarse junto con 'from'."
          },
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Zona horaria IANA utilizada para formatear la columna 'hour'. Por defecto: America/Bogota."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2000
            },
            "description": "Cantidad máxima de filas por página. Por defecto 500."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Número de página a recuperar. Por defecto 1."
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado de métricas horarias.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HourlyReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos o rango incorrecto.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/daily": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Construye el reporte diario de 24 horas",
        "description": "Ensambla un resumen con 24 filas (una por hora) y métricas globales del día solicitado: extremos de temperatura, humedad promedio y energía solar. Requiere deviceId y date (YYYY-MM-DD) y respeta la zona horaria solicitada para formatear la respuesta.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identificador del dispositivo que se desea consultar."
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "description": "Día a consultar en formato YYYY-MM-DD."
          },
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Zona horaria para interpretar y formatear el día. Por defecto America/Bogota."
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte diario generado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/monthly": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Construye el reporte mensual por dispositivo",
        "description": "Devuelve una lista con un registro por día del mes solicitado, incluyendo temperaturas máximas/mínimas/promedio, humedad relativa y radiación total, promedio y máxima. Sirve para componentes de calendario o tablas de seguimiento mensual.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identificador del dispositivo que se desea consultar."
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 2000,
              "maximum": 2100
            },
            "description": "Año en formato YYYY."
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            },
            "description": "Mes numérico (1-12)."
          },
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Zona horaria para interpretar el mes. Por defecto America/Bogota."
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte mensual generado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/hourly/recalculate": {
      "post": {
        "tags": [
          "Reportes"
        ],
        "summary": "Reprocesa los promedios horarios en un rango específico",
        "description": "Ejecuta nuevamente la canalización de agregación sobre lecturas crudas en el rango [from, to), opcionalmente filtrando por deviceId y/o sensorType. Requiere permisos de administrador (salvo que REPORTS_AUTH_DISABLED sea true) y responde con el detalle de filas recalculadas para que el frontend confirme la acción al operador.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HourlyRecalculateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proceso completado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Agregados horarios recalculados correctamente."
                    },
                    "upserted": {
                      "type": "integer",
                      "example": 48
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos o rango incorrecto.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Incluye el token JWT obtenido al iniciar sesión en el encabezado `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Cookie HTTP con el refresh token emitido en el inicio de sesión."
      }
    },
    "schemas": {
...
      "HourlyReportEntry": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "example": "ESP32_1"
          },
          "sensorType": {
            "type": "string",
            "enum": [
              "temperature",
              "humidity",
              "soil_humidity",
              "solar_radiation"
            ],
            "example": "temperature"
          },
          "hour": {
            "type": "string",
            "example": "2025-11-02T09:00:00-05:00"
          },
          "avg": {
            "type": "number",
            "example": 22.4
          },
          "min": {
            "type": "number",
            "example": 20.1
          },
          "max": {
            "type": "number",
            "example": 24.9
          },
          "samples": {
            "type": "integer",
            "example": 12
          },
          "units": {
            "type": "string",
            "example": "°C"
          }
        }
      },
      "HourlyReportPagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 96
          },
          "limit": {
            "type": "integer",
            "example": 50
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "pages": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "HourlyReportResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HourlyReportEntry"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/HourlyReportPagination"
          }
        },
        "example": {
          "data": [
            {
              "deviceId": "ESP32_1",
              "sensorType": "temperature",
              "hour": "2025-11-06T12:00:00-05:00",
              "avg": 29.9,
              "min": 10.47,
              "max": 49.87,
              "samples": 111,
              "units": "°C"
            }
          ],
          "pagination": {
            "total": 1,
            "limit": 500,
            "page": 1,
            "pages": 1
          }
        }
      },
      "DailyReportRow": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "minimum": 0,
            "maximum": 23,
            "example": 13
          },
          "solar_radiation_avg": {
            "type": "number",
            "nullable": true,
            "example": 520
          },
          "humidity_avg": {
            "type": "number",
            "nullable": true,
            "example": 68.4
          },
          "temperature_avg": {
            "type": "number",
            "nullable": true,
            "example": 24.7
          },
          "isTmax": {
            "type": "boolean",
            "nullable": true,
            "example": true
          },
          "isTmin": {
            "type": "boolean",
            "nullable": true,
            "example": false
          }
        }
      },
      "DailyReportTemperatureSummary": {
        "type": "object",
        "properties": {
          "tmax": {
            "type": "number",
            "nullable": true,
            "example": 30.5
          },
          "tmin": {
            "type": "number",
            "nullable": true,
            "example": 18.2
          },
          "tpro": {
            "type": "number",
            "nullable": true,
            "example": 24.6
          }
        }
      },
      "DailyReportHumiditySummary": {
        "type": "object",
        "properties": {
          "hpro": {
            "type": "number",
            "nullable": true,
            "example": 63.2
          }
        }
      },
      "DailyReportRadiationSummary": {
        "type": "object",
        "properties": {
          "radTot": {
            "type": "number",
            "nullable": true,
            "example": 4860
          },
          "radPro": {
            "type": "number",
            "nullable": true,
            "example": 405
          },
          "radMax": {
            "type": "number",
            "nullable": true,
            "example": 710
          }
        }
      },
      "DailyReportResponse": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "example": "ESP32_1"
          },
          "date": {
            "type": "string",
            "example": "2025-11-02"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportRow"
            }
          },
          "temperature": {
            "$ref": "#/components/schemas/DailyReportTemperatureSummary"
          },
          "humidity": {
            "$ref": "#/components/schemas/DailyReportHumiditySummary"
          },
          "radiation": {
            "$ref": "#/components/schemas/DailyReportRadiationSummary"
          }
        },
        "example": {
          "deviceId": "ESP32_1",
          "date": "2025-11-06",
          "rows": [
            {
              "hour": 0
            },
            {
              "hour": 1
            },
            {
              "hour": 2
            },
            {
              "hour": 3
            },
            {
              "hour": 4
            },
            {
              "hour": 5
            },
            {
              "hour": 6
            },
            {
              "hour": 7
            },
            {
              "hour": 8
            },
            {
              "hour": 9
            },
            {
              "hour": 10
            },
            {
              "hour": 11
            },
            {
              "hour": 12,
              "humidity_avg": 78.65,
              "solar_radiation_avg": 626.23,
              "temperature_avg": 29.9,
              "isTmax": true,
              "isTmin": true
            },
            {
              "hour": 13
            },
            {
              "hour": 14
            },
            {
              "hour": 15
            },
            {
              "hour": 16
            },
            {
              "hour": 17
            },
            {
              "hour": 18
            },
            {
              "hour": 19
            },
            {
              "hour": 20
            },
            {
              "hour": 21
            },
            {
              "hour": 22
            },
            {
              "hour": 23
            }
          ],
          "temperature": {
            "tmax": 29.9,
            "tmin": 29.9,
            "tpro": 29.9
          },
          "humidity": {
            "hpro": 78.65
          },
          "radiation": {
            "radTot": 626.23,
            "radPro": 626.23,
            "radMax": 626.23
          }
        }
      },
      "MonthlyReportDay": {
        "type": "object",
        "properties": {
          "day": {
            "type": "integer",
            "minimum": 1,
            "maximum": 31,
            "example": 12
          },
          "RadTot": {
            "type": "number",
            "nullable": true,
            "example": 480
          },
          "RadPro": {
            "type": "number",
            "nullable": true,
            "example": 320
          },
          "RadMax": {
            "type": "number",
            "nullable": true,
            "example": 650
          },
          "HR": {
            "type": "number",
            "nullable": true,
            "example": 65.3
          },
          "Tmax": {
            "type": "number",
            "nullable": true,
            "example": 29.4
          },
          "Tmin": {
            "type": "number",
            "nullable": true,
            "example": 17.6
          },
          "Tpro": {
            "type": "number",
            "nullable": true,
            "example": 23.1
          }
        }
      },
      "MonthlyReportResponse": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "example": "ESP32_1"
          },
          "year": {
            "type": "integer",
            "example": 2025
          },
          "month": {
            "type": "integer",
            "example": 11
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyReportDay"
            }
          }
        },
        "example": {
          "deviceId": "ESP32_1",
          "year": 2025,
          "month": 11,
          "days": [
            {
              "day": 1
            },
            {
              "day": 2
            },
            {
              "day": 3
            },
            {
              "day": 4
            },
            {
              "day": 5
            },
            {
              "day": 6,
              "RadTot": 626.23,
              "RadPro": 626.23,
              "RadMax": 626.23,
              "HR": 78.65,
              "Tmax": 29.9,
              "Tmin": 29.9,
              "Tpro": 29.9
            },
            {
              "day": 7
            },
            {
              "day": 8
            },
            {
              "day": 9
            },
            {
              "day": 10
            },
            {
              "day": 11
            },
            {
              "day": 12
            },
            {
              "day": 13
            },
            {
              "day": 14
            },
            {
              "day": 15
            },
            {
              "day": 16
            },
            {
              "day": 17
            },
            {
              "day": 18
            },
            {
              "day": 19
            },
            {
              "day": 20
            },
            {
              "day": 21
            },
            {
              "day": 22
            },
            {
              "day": 23
            },
            {
              "day": 24
            },
            {
              "day": 25
            },
            {
              "day": 26
            },
            {
              "day": 27
            },
            {
              "day": 28
            },
            {
              "day": 29
            },
            {
              "day": 30
            }
          ]
        }
      },
      "HourlyRecalculateRequest": {
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "properties": {
          "deviceId": {
            "type": "string",
            "nullable": true,
            "example": "ESP32_1"
          },
          "sensorType": {
            "type": "string",
            "enum": [
              "temperature",
              "humidity",
              "soil_humidity",
              "solar_radiation"
            ],
            "nullable": true,
            "example": "temperature"
          },
          "from": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-02T00:00:00-05:00"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-03T00:00:00-05:00"
          },
          "timezone": {
            "type": "string",
            "nullable": true,
            "example": "America/Bogota"
          }
        },
        "example": {
          "deviceId": "ESP32_1",
          "sensorType": "temperature",
          "from": "2025-11-02T00:00:00-05:00",
          "to": "2025-11-03T00:00:00-05:00",
          "timezone": "America/Bogota"
        }
      },
...
      "ErrorRespuesta": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "No se pudo completar la operación."
          }
        }
      }
    }
  }
}